+++ packages/eslint-config-react/index.js
@@ -8,10 +8,10 @@ module.exports = {
     browser: true,
   },
   extends: [
-    './rules/plugin-react',
-    './rules/plugin-react-jsx',
-    './rules/plugin-react-hooks',
-    './rules/plugin-jsx-a11y',
+    './core/plugin-react',
+    './core/plugin-react-jsx',
+    './core/plugin-react-hooks',
+    './core/plugin-jsx-a11y',
   ].map(require.resolve),
   parser: '@babel/eslint-parser',
   parserOptions: {