+++ packages/eslint-config-react/CHANGELOG.md
@@ -21,6 +21,8 @@
   * There are some limitations with this rule (forwardRefs) which make it problematic at times.
   * defaultProps is being phased out in favor of destructured default values in function components making this rule have lesser value than before.
 * Replace `babel-eslint` with `@babel/eslint-parser`.
+* Restructure ruleset directories so there's distinction between what is core to the config and what is considered "optional" or "addons"
+
 
 #### Fixed
 * resolved an issue where `indent` and `react/jsx-indent` conflict.